set up build workflow
Create a new workflow for building Rust to a wasm project. It is assumed
that rust is installed by default on the worker, however it may be
advisable to explicitly install it in the future. This is meant to be an
initial pass of a build pipeline which can be elaborated on in the
future.