add deploy job for github pages
Amend the original build job to publish the result as an artifact and
create a new job which publishes the resulting artifact to GitHub Pages.
This uses some beta and alpha features which should not be relied upon
until they reach general availability, but they should be sufficient for
this initial development period.